Move test back into separate file
